Add files via upload
I had a bug in the previous commit that I couldn't find, so I rewrote all the code to re-follow my logic and find the typo. I also resized the mermaid to better fit the screen, changed the bg colour, and made it so the mermaid couldn't move off the screen (i.e. set a range from 0 to self.screen_rect.right()).