Add files via upload
The mermaid can fire bubbles now! I made a bubble class where the bubble is a blue 5 px by 5 px square that starts from the top of the mermaid.rect. The speed of the bubble moves slower than the mermaid. I also made it so the player can only fire three bullets at a time (three bullets on the screen at a time), and the bullet counter goes back down when the bullet leaves the screen. I finally assigned all of the bubble functions as their own individual function to simplify the overall game while loop. LOTS OF PROGRESS <3